Refactor onResize and onVisibleChange to DomUtils
