+++ utils/react/Hooks.js
@@ -6,8 +6,8 @@
  */
 /* eslint-disable react-hooks/exhaustive-deps */
 import {useCallback, useEffect} from 'react';
-import ResizeObserver from 'resize-observer-polyfill';
-import {isFinite, isNil, debounce as lodashDebounce} from 'lodash';
+import {isNil} from 'lodash';
+import {onResize, onVisibleChange} from '@xh/hoist/utils/js';
 
 /**
  * Hook to run a function once after component has been mounted.
@@ -33,14 +33,13 @@ export function useOnUnmount(fn) {
 
 /**
  * Hook to run a function when component is resized or content becomes visible/invisible.
- * The resize callback will not run the hook when the size is changed to 0 or is changed back from 0 to a previous
- * size. This is to improve performance by avoiding unneeded resizing.
  * @param {function} fn
  * @param {Object} [c] - configuration object
  * @param {number} [c.dimsDebounce] - milliseconds to debounce
  * @param {function} [c.fnDims] - called on width or height change
  * @param {function} [c.fnVis] - called when visible change
- * @returns {Ref} - ref to be placed on target component
+ * @returns {Ref} - callback ref to be placed on target component - this is not
+ *                  a DOM node.  Its the callback passed to `useCallBack`
  */
 export function useOnVisDimsChange({fnDims, fnVis, dimsDebounce}) {
     let node, dimsObserver, visObserver;
@@ -48,7 +47,7 @@ export function useOnVisDimsChange({fnDims, fnVis, dimsDebounce}) {
     // see React example: https://reactjs.org/docs/hooks-faq.html#how-can-i-measure-a-dom-node
     const ret = useCallback(_node => {
         if (!isNil(_node)) {
-            dimsObserver = onResize(fnDims, dimsDebounce, _node);
+            dimsObserver = onResize(fnDims, _node, {debounce: dimsDebounce});
             visObserver = onVisibleChange(fnVis, _node);
             node = _node;
         }
@@ -66,8 +65,6 @@ export function useOnVisDimsChange({fnDims, fnVis, dimsDebounce}) {
 
 /**
  * Hook to run a function when component is resized.
- * This will not run the hook when the size is changed to 0 or is changed back from 0 to a previous
- * size. This is to improve performance by avoiding unneeded resizing.
  * @param {function} fn
  * @param {Object} [c] - configuration object
  * @param {number} [c.debounce] - milliseconds to debounce
@@ -75,11 +72,9 @@ export function useOnVisDimsChange({fnDims, fnVis, dimsDebounce}) {
  */
 export function useOnResize(fn, {debounce} = {}) {
     let node, resizeObserver;
-
-    // see React example: https://reactjs.org/docs/hooks-faq.html#how-can-i-measure-a-dom-node
     const ret = useCallback(_node => {
         if (!isNil(_node)) {
-            resizeObserver = onResize(fn, debounce, _node);
+            resizeObserver = onResize(fn, _node, {debounce});
             node = _node;
         }
     });
@@ -111,49 +106,4 @@ export function useOnVisibleChange(fn) {
     }, [node]);
 
     return ret;
-}
-
-
-function onResize(fn, debounce, node) {
-    if (fn) {
-        let prevWidth, prevHeight;
-        const wrappedFn = (e) => {
-            const {width, height} = e[0].contentRect,
-                isVisible = width !== 0 && height !== 0,
-                hasChanged = width !== prevWidth || height !== prevHeight;
-
-            if (isVisible && hasChanged) {
-                prevWidth = width;
-                prevHeight = height;
-                fn(e);
-            }
-        };
-
-        const callbackFn = isFinite(debounce) && debounce >= 0 ?
-                lodashDebounce(wrappedFn, debounce) : wrappedFn,
-            observer = new ResizeObserver(callbackFn);
-
-        observer.observe(node);
-        return observer;
-    }
-}
-
-function onVisibleChange(fn, node) {
-    if (fn) {
-        let prevVisible;
-        const observer = new ResizeObserver((e) => {
-            const {width, height} = e[0].contentRect,
-                visible = width !== 0 && height !== 0,
-                hasChanged = visible !== prevVisible;
-
-            if (hasChanged) {
-                prevVisible = visible;
-                fn(visible);
-            }
-        });
-
-        observer.observe(node);
-        return observer;
-    }
-
 }
